requirements.txt
beautifulsoup4==4.12.3
bitarray==2.3.5
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
debugpy==1.6.6
Deprecated==1.2.14
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
lxml==5.1.0
pycparser==2.21
PyGithub==1.57
PyJWT==2.8.0
PyNaCl==1.5.0
python-magic==0.4.24
pymongo==4.6.2
PyYAML==6.0.1
requests==2.31.0
retrying==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
urllib3==2.1.0
wrapt==1.16.0